Added a popup ui widget and (temporary) graphics. The widget supports resizing and multiple popup windows can be opened at the same time.
Fixed layout issues with Slider widget.

Fixed layout issues when centering text.

Fixed changing the visiblity of the menu scrollbar when the window is 
resized.

Improved the performance of textarea. The textarea now correctly decides 
if a scrollbar is needed.

Fixed event handling. Applets adding event handlers should be able to 
override the default event actions.



git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@128 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9