Fixed crasher in textarea with scrollbars.
Fixed crasher with stopped timers.

Refactored the wallpaper setting into a seperate applet. Removed a couple of the 
wallpapers that don't render well. Default to a black background if one is not set.

Don't draw the iconbar background on the wallpaper when it's loaded, render it in the 
global widgets.



git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@181 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9