* This checkin breaks the applet API *
Allow applets to store settings. Implemented for setup lanaguage, screensavers, 
flick screensaver and wallpaper applets.

Currently the settings are stored in the same directory as the applet code, this 
is temporary and will change.

API changes are:
* AppletManager:load() renamed to AppletManager:loadApplet(). This is matches the 
other applet manager calls, and saves confusion with the lua load function.
* Applets should implement the init() method, not the __init() object constructor. 
The applets strings and settings are available in the init() method.
* defaultSettings() is now an AppletMeta function, not an Applet function.





git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@200 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9