Description: Remove an inconvenient assertion.
Hierarchical settings menus sent from SS are cached. The radio associated with each item is
cached as well. When entering, leaving and entering again such a menu, the radios are 
re-used while the window, menus and labels are new.

This assertion prevented this behaviour, despite it being harmless (because anything
having a link forward to any of these radios was no longer viewable and candidate for GC).

The alternative was putting in place either some code in the SB menu walking machine to 
forget about all radios when closing a window OR some scaffolding in the UI classes to
remove any leftover forward links.
Both options were considered a bit expensive and complicated.

So it was decided to remove the assertion and see about any problems.



git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@323 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9