Allow the user to select the player at the end of setup.
Fix bug where multiple failure windows sometimes opened if we could not 
connect to the network.

Backoff gracefully when polling for slimservers, up to a 30 seconds 
interval.

Fixed a timer bug, the timer was not restarted correctly if the interval 
was changed. Also reverted r334 as that fix was wrong.

Bug: #5219
Tell a white lie about WEP, claim it is encryption :).




git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@339 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9