Description: As you scroll through languages on first power up, the hint bar at the bottom should change languages.
Implemented focus events on menu items, and focus callbacks for the SimpleMenu. This is used to update the language as you scroll through the langauge options in the setup/settings menu. In addition this allow the wallpaper to be previewed in the wallpaper applet.

The focus events will also be used to allow long lines to scroll (Bug #5243).


Bug: #5250

git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@349 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9