Description: Updated click sound.
Bug: N/A
Description: Change how frame rate is calculated in the event loop. This fixes when the timer wraps around (thanks triode), and also keeps some 
time for event processing when we can't maintain the desired framerate. This should improve the overall system throughtput.

Bug: N/A
Description: Added some debug to the Network Thread to try to discover why the pump is sometimes nil.

Bug: 5518
Description: Add lua garbage collection into the main event loop. This should manage the heap better, and act as a work around to the 
screensavers crashing the jive remote.



git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@524 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9