Description: Implement FWD button to skip to next page in setup. Fixed, can use FWD or REW on any page during setup.
Note that FWD/REW have been added as forward/backward actions for menus. To allow 
SlimBrowser to override these for playback control the menu handler has been changed to 
allow the selected widget to process the key event before the standard menu key handling. 
This change does not appear to have broken the existing functionality, and it will allow 
more flexiblity for applets to override the default behaviour.


Bug: 5456

git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@806 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9