Screensaver rework. The original screensaver applet assumed that a screensaver would only open a single window, in practice this has not been the case. This resulted in bugs where the screensaver could start multiple times, and sometimes it was difficult to stop the screensaver. A modified screensaver api now means that the screensavers need to register a window using ScreenSavers:screensaverWindow(). This installs some listeners to manage the screensaver state, and provides a defaultkey/scroll listener to close the screensaver.
Clock screensaver: code tidy up and transitions.

Flickr screensaver: pressing Go lets you read the users comments on the picture. This is to demonstrate how a screensaver can have 
sub-windows (the now playing screensaver will need this). Don't do dns in the main thread, that could cause the ui to block.

Now playing screensaver: code tidy up and transitions. Fixed no album images.

Fixed a bug in widget where the timers would start even when the widget was not displayed.

Renamed and updated documentation to make it clear that the tcp sockets can take an ip address or hostname.



git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@917 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9