Player object stores all the playerstatus data in playerStatus table NowPlaying access all data from player object instead of managing it itself
Only make trackPos correction against lastSeen if mode = 'play'

Small change to SlimBrowser in free(). I'm seeing Jive on the desktop hang occasionally when when unsubscribing to menustatus for a player. Not sure if this addresses the issue, but put in a check to make sure _server and _player exist before unsubscribing.

Bug: 6265

git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@1344 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9