Decription: Handle custom navigation 'nextWindow' in SlimBrowser to allow navigation to "special" windows, nowPlaying and home This will allow certain go actions (e.g., starting a random mix) to automatically push the user to the NowPlaying or Home screen
Added local functions _goHome and _goNowPlaying that are called in a number of spots for both nextWindow actions and global key handlers

Re-add the handling of a special 'READ ME' item that may get sent over by SC that should be discared by this revision of Jive.
This is probably not necessary, but want to make sure we don't get bit during the upgrade path of customers.

Bug: 6304

git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@1406 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9