Fix behaviour of popup windows. This adds some new properites to the Window: - setAlwaysOnTop() to force this window to the top of the window stack. - setShowFrameworkWidgets() to enable/disable rendering of global widgets. - setTransparent() to allows the lower window to be active and redrawn.
Redefined Popup windows to, by default, not allow screensavers, always be on top, auto 
hide, not show global widgets and to be transparent.

This corrects the behaviour of popup widgets, and in particular fixes a rendering problem 
when using transparent popups. This does not yet correctly render window transitions 
while popups are visible.


Bug: 5387

git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/jive@1426 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9