Description: though significant effort went into these lines, in chunked responses, we still weren't getting the return value right in all instances.
In the case of when your list count is exactly equal to the BLOCK_SIZE * lastKey, it is necessary to decrement lastKey by 1 or you get in an endless request loop. It looks fine on the UI, but in the background it just keep requesting and requesting the last chunk.

Bug: 6657

git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/squeezeplay@2143 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9