Description: initial pass at CustomizeHomeMenuApplet.
Note: this isn't fully functional yet, particularly with regard to reading existing settings at boot-time and maintaining that data between changing players. Also, there need to be some special case assessments for particular items, e.g., should you be able to hide the customize home menu item, effectively eliminating your ability to change anything without a factory reset.

It does the basics correctly, however: allows a user to move an item from it's default node to the home node, or to hide it from view.

I have not added the applet to the makefile yet, so the applet will not be exposed to the UI yet.

Bug: 5366

git-svn-id: https://svn.slimdevices.com/repos/jive/7.1/trunk/squeezeplay@2416 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9