IMPORTANT: Made the jnt subscriptions weak, so if the function subscribing goes out of scope then the subscription is automatically cancelled.
All code the subscribes using the jnt needs reviewing to see if it breaks.

This change is needed to make sure objects like SlimServer get correctly garbage collected 
when on longer required (in this case when the SC is stopped). The jnt subscription was 
keeping the object alive in the garbage collector.

This change will cause problems if you subscribe a closure and don't keep a reference to 
the closure in the calling code. I decided that the new behaviour is correct, because 
previously any unreferenced closure actually created a memory leak.


Bug: 6683

git-svn-id: https://svn.slimdevices.com/repos/jive/7.1/branches/discovery-refactor/squeezeplay@2558 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9