Decode to PCM even when paused. This fixes a couple of bugs, an audio glitch when playback starts and avoiding an auto-pause (that never resumed) on an audio underrun.
A delay is still needed before the decoder starts, to prevent immediate decoder underruns at the start of a track. Also (at least on 
ip3k) some decoders assume sufficient data to parse the headers at the start of a track.

The auto start threshold/STMl has been changed to trigger on the bytes received, not the fullness of the decode buffer. A future 
improvement maybe to trigger on the fullness of the output buffer, but Alan says that this will require a SC change and should be 
done in all the player software (squeezeplay, ip3k, etc) at the sametime.


Bug: 6442

git-svn-id: https://svn.slimdevices.com/repos/jive/7.3/trunk/squeezeplay@3307 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9