Checkpoint work in progress. This gets as far as Connecting to.. in both the wlan and eth cases.
Clean up the flow in SetupWelcomeApplet. This removes code only needed on jive for connecting 
via ad-hoc. Also it passes the choosen interface (wlan or eth) through other functions.
Refactor SetupNetworkingApplet. Most functions now are modified to pass the interface and ssid 
in as arguments. Made the eth case be handled via the same flow.



git-svn-id: https://svn.slimdevices.com/repos/jive/7.4/private-branches/fab4/squeezeplay@3831 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9