Bug #12889, Bug #12753.  I implemented the lower brightness mode.  This allows the baby display to go very dim.
I had to change the applet code so that it uses the bl_power in the /sys interface to switch to dim mode.  This required that each entry of the brightnessTable contains a pair of values, rather than a single value.  

I believe I got all the manual settings, but I think I broke the ambient light sensor code (which was kind of broken anyway).  

Raphael, would you mind taking a look at that, and see if you can see what I broke?

The brightness is always an *index* into brightnessTable, rather than a value.  This means that you don't want to be multiplying brightnesses, only adding and subtracting. 

Only setBrightness should dig into the brightnessTable values.

-Caleb


git-svn-id: https://svn.slimdevices.com/repos/jive/7.4/private-branches/fab4/squeezeplay@7165 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9