Description: remove bottom up window transition on showBrieflies.
This may be as good as it gets for 7.4...I've specified the window transitions to be FadeIn and FadeOut, but they don't actually work because the showBriefly is not the size of the window below it, which causes the transition effect to not happen. So it just appears immediately on screen without a fade effect. Slide up from the bottom is gone though.
Bug: 13725

git-svn-id: https://svn.slimdevices.com/repos/jive/7.4/private-branches/fab4/squeezeplay@7288 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9