Fixed looping tracks in SP. This problem did not always happen, but appears to be more of a problem on slower devices.
The streambuf loop pointer was often being set incorrectly, due to a difference in the network stacks on SP 
and ip3k. To fix this the loop pointer is now always set after the http headers. If looping is then required 
the loop flag is flipped to turn it on.

This removes the streambuf_*_loop public interface. The semantics have changed, and this interface was not 
actually used.

Tested with the backup alarm as described in the bug.


Bug #13993

git-svn-id: https://svn.slimdevices.com/repos/jive/7.4/trunk/squeezeplay@7681 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9