New progress group widgets were being added on every track change but not cleared, in a bogus swap out situation.  Now that we hold onto the np window rather than replacing it, this leak eventually caused heavy cpu load.
Note: This also fixed the issue that remote streams shouldn't show the progress dialog graphic, but did.

Fix is to correctly swap the two progress bars.

Also, moved the _updatePosition timers being used on each of the progress groups and the progress slider groupto a window timer, to avoid double calls to updatePosition()
Fixed Bug: 14162

git-svn-id: https://svn.slimdevices.com/repos/jive/7.4/trunk/squeezeplay@7821 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9