Description: when playlist size is 1, gray out the fwd and shuffle buttons when playlist size is 1 and track has no duration (i.e., internet radio), gray out the rew button
the repeat button has validity in all cases and will not be grayed out. it will force a restart of the stream if for some reason it stops during playback.

fwd, rew, and shuffle buttons all have validity for internet radio when the current playlist size is > 1, so in that case they stay. This is a corner case, but the typical user will never see it as the common case is to play an internet radio stream as a single item playlist.

Fixed Bug: 15085

git-svn-id: https://svn.slimdevices.com/repos/jive/7.5/trunk/squeezeplay@8120 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9