Description: make RTC fallback alarm running at all times instead of only when server is detected as disconnected assumption is that notify_serverDisconnected and notify_serverConnected are not always trustworthy sources of information
when alarm time hits and server is connected, two things should happen-- RTC timer should timeout and attempt to fire the alarm, and server should send a notification that the alarm is active and attempt to fire the alarm. Whichever one loses that race will simply not do anything. Whichever one wins that race will see if the player is connected and fire the appropriate alarm (if connected, audio from server/network, if not, local fallback alarm)

this bug is not fixed, but is the first in hopefully a series of checkins between now and xmas to harden this feature.

Bug: 14870

git-svn-id: https://svn.slimdevices.com/repos/jive/7.4/trunk/squeezeplay@8237 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9