Fixed Bug: 10064 Description: move playerstatus subscription to a "full push" model rather than push + 30 second poll
the danger here is if playerstatus isn't auto-generated when it should be (bug 10064) and instead of refreshing withing 30 seconds, it's stale until the next notification auto-generate, which could be much longer.

It's time to quit evading this bug though-- if 10064 still is out there, let's flush out those bugs and get them fixed rather than take this defensive stand that makes the network chatter much higher than it should be.

Fixed Bug: 8300

git-svn-id: https://svn.slimdevices.com/repos/jive/7.4/trunk/squeezeplay@8262 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9