Fixed bug 16617: Controller (SqueezePlay) may never reconnect fully to MySB.com after connection loss
Fix incorrect use of self.state (should be oldstate) in setEndpoint(). Not actually provoking a bug at present because the only call to setEndpoint() is bracketed by disconnect()/connect() calls.

Server may send reconnection advice of reconnect=none under certain error conditions. This causes Comet.lua to enter UNCONNECTED state and not attempt a retry, but SlimServer.lua will assume that it is CONNECTING state (retrying). In order for it to be possible to escape from this situation, in the case that whatever provoked the reconnect=none was actually a transient condition, we allow a new request to force a reconnection attempt.

When re-enabling subscriptions that were previously active as part of a new handshake, also remove them from the set of sent requests that have not yet been acknowledged, and that will get resent when the new connection is established; otherwise duplicate subscription requests can be sent.

The RE used to match clientId strings in the response channels of subscription requests in the queue was insufficient. It needs also to match 'X' as well as hexadecimal characters. 

Include /meta/subscribe along with /meta/reconnect in case this reconnect is happening after a successful handshake but failed /meta/connect.

git-svn-id: https://svn.slimdevices.com/repos/jive/7.6/trunk/squeezeplay@9209 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9