Task  B0020: Player control over UDP
Extend UDAP with pause, set_volume and get_ip methods that will be acted upon if a local player exists (regardless of configuration state).
Extend Slimproto audg to carry controller-id in addition to sequence number. Sequence numbers associated with a controller-id are from a disjoint space with normal sequence numbers and are used only to reconcile the order of received volume commands that might come via UDA or Slimproto.


git-svn-id: https://svn.slimdevices.com/repos/jive/7.8/trunk/squeezeplay@9677 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9