Fixed bug 17758: Pop up symbols linger too long on the screen
Do not set a duration of 10000 on jive show-brieflies for current song state, which are used for playlist jump events, just to ensure that the new track title is pushed to the first line of the display and stays there until a new playerStatus arrives (which can take quite a long time under some circumstances). This affected both the duration of the new title on the display and that of the the Play pop-up icon if the command was invoked via IR.  Having the popup hang around for 10s can be irritating. Given that the player has control of title change itself, it can set a long duration on just that element and use the default show-briefly duration for the popup.


git-svn-id: https://svn.slimdevices.com/repos/jive/7.8/trunk/squeezeplay@9683 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9