Update context-menu styles to use icons for playlist control actions.
Use play and add icons for Play, Add, Insert and PlayAll actions (styles), and a heart icon for item_fav style.
This are particularly relevant now that defeat-touch-to-play is the default server setting.

Server change are needed send the appropriate styles but I will wait a day or so before checking those in so that f/w for players is available first.


git-svn-id: https://svn.slimdevices.com/repos/jive/7.8/trunk/squeezeplay@9703 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9