Allow switching between different SN
Added functionality to download config JSON file from the config server during boot-time or during initial device setup (OOB).
Depending on the content of the JSON file the device uses a specific SN instance (only one option returned)
or if more than one option is present in the JSON file an additional menu 'Developer' is added in the 'Advanced'
settings which allows to switch between different SN instances on the fly.

Change-Id: I882091e29210f8c51b6d38bb504384d8e699287b