More on browsing attached servers: actions with nextWindow=parent...
The handling of most actions with a nextWindow option needs to create
a step (but not necessarily a sink) so that the step.server can be
based on the currentStep server. In this case, the step is only being
created so that it has a server and the other inferred semantics of
_pushToNewWindow(step), that were previously there by default, is not
appropriate.

This change reverses the default state of skipNewWindowPush flag,
which controls calling of _pushToNewWindow(step), and sets as
appropriate in each of the cases.

Change-Id: Ifc1e5878159f5e9403a1de906b1ec7e07d8ed46d