Pass controlling server address in Comet handshake to other servers
The UEML server code needs to know the address (IP + TCP port)
of its controlling (SN) server so that playlist modification
requests can be sent to the right place. To facilitate this,
SlimServer.lua informs its attached Comet connection object of
this address, both initially and whenever the address or
controlling server changes. If the address changes while Comet
is connected a new handskae is forced.

The force of a new hanshake exposed a bug whereby simple requests
over Comet that were expecting a response could get dropped and
not resent on the re-handshaked connection if the acknowledgement
arrived but the actual data response did not. This is resolved
by only using the actual data response to remove a pending
request from the sent_reqs array.

Change-Id: Ie46258a39cd38c7f1b5b0555192cab5d7f06bb59