Player-status reliability: improve re-subscription handling
Add pending unsubscribe requests to request packets before pending subscription requests in case a timing window with network reconnects results in the possibility of an old unsubscribe negating a newer subscribe.

Only remove requests from pending queue if an error did not result in _handleAdvice().

Only mark a player offStage if it is already onStage. This was the cause of unnecessary unsubscribe requests.