Refactor button-remapping code.
Look for button remapping in playerStatus.buttons instead of playerStatus.remoteMeta.buttons (the old location is maintained for backwards compatibility).

Explicitly use a button remapping of type 'like' for the repeat button and one of 'dislike' for the shuffle button (the old names, 'repeat' & 'shuffle' are used as a fallback for backwards compatibility).