Bugfix: Incomplete extraction of icy metadata causes Squeezeplay to quit From Martin Williams <martinr.williams@gmail.com>
Extracting icy metadata from an incoming stream requires two data reads
from the fifo circular stream buffer if the icy metadata should roll
around the end of the buffer.

At present only one read is attempted, and the code assumes that a short
read implies a fundamental failure in the integrity of the stream
buffer. [assert(r == icy_len)].

The consequence of the assertion failure is that Squeezeplay will simply
quit.

This change corrects matters by undertaking two reads if the initial
read is 'short'. This behaviour is mandated by the implementation of
'streambuf_fast_read' and the fifo circular stream buffer.

We know that there are sufficient bytes available to satisfy the full
read. So the only valid state after the first 'short' read is that the
fifo read buffer pointer is set at the start of the buffer. We implement
an assertion check on that condition. It assures integrity of the stream
buffer, and is in the spirit of the existing code.

The issue surfaced from the Squeezebox forum thread "Squeezebox Radios
Reboots, with talk radio in the UK, cable connected , cable connect".

https://forums.slimdevices.com/showthread.php?113309-Squeezebox-Radios-Reboots-with-talk-radio-in-the-UK-cable-connected-cable-connect

Here, users of the UK based LBC & Classic FM radio streams noted that
their Squeezebox Radios would spontaneously reboot when playing these
streams for relatively long periods (hours). But the issue is not
confined to these streams. It could occur for any stream that has been
running long enough, and where icy metadata just happens to roll around
the end of the fifo buffer. The odds of a reboot increase proportionally
with length of time that a stream is continuously played, and with the
length of metadata that the stream typically provides.

Thanks mw9.

Increase squeezeplay revision to 1310.