Bugfix: Wi-fi scan on Radio fails under later versions of wpa_supplicant From Martin Williams <martinr.williams@gmail.com>
On a Radio, '_wirelessScanTask' in 'net/Networking.lua' fails to
enumerate the available wi-fi networks under later versions of
wpa_supplicant.

This is due to a change in the way that wpa_supplicant's wext driver
reports signal strength. In consequence, '_wirelessScanTask' simply
won't match any available networks, and the user is forced to enter the
SSID manually.

The Radio uses the wext driver, and is impacted. At time of writing,
an updated wpa_supplicant is only provided for the Radio, because the
the Touch and the Controller are both using an earlier version of
wpa_supplicant (0.5.7) with a closed source marvell driver.

This change addresses the issue by adapting '_wirelessScanTask' to
accommodate both "old style" and "new style" signal strength reporting.
It is a "no-op" for systems using "old style" reporting (wpa_supplicant
<= (approx.) 0.7).

Further remarks:

On or around 4 October 2010 the wext driver to wpa_supplicant was
corrected to report the signal level in dBm when the hardware/kernel
driver indicated so. This change was made subsequent to release 0.7.2,
but before release 2.0. The Radio uses the wext driver (option -Dwext to
wpa_supplement).

Refer 'wext_get_scan_qual' in 'drivers/driver_wext.c'.
Commit 32f4e7 in git://w1.fi/srv/git/hostap.git:
wext: Fix scan result signal levels when driver reports in dBm

The effect is that the signal level returned from wpa_supplicant's wi-fi
scan may now contain negative numbers. The allowed range is -192dBm to
+63dBm.

Importantly, we know that we can get back the number "we first started
with" ("old style") by simply adding back 0x100 if the number is
negative. That will simply undo what the later versions of the
wpa_supplicant wext driver will have done.

So that is what we do. '_wirelessScanTask' will now accommodate both
styles of signal strength reporting.

The point may need to be re-examined if/when wpa_supplicant is updated
on the Touch and Controller platforms.

Some further background explanation here:
https://sources.debian.org/src/wireless-tools/30~pre9-13/iwlib.c/#L1492