Merge pull request #14: remove -force_cpusubtype_ALL gcc flag from squeezeplay Makefile.
The gcc flag -force_cpusubtype_ALL appears to have disappeared from Apple Clang 15.0.0. After a cursory look at the v15 release notes and source, I could not see any obvious references to it, so perhaps it is an Apple thing.

This flag is ppc64 related. Although it does show up in SDL-1.2.15, it is in $EXTRA_CFLAGS which its Makefile never uses, so it is harmless in that directory.

Increase squeezeplay revision to 1451.
Thanks kerneljake.